+++ tests/test_eyconf_extra_fields.py
@@ -65,6 +65,16 @@ class TestEYConfExtraFields:
             "str_field": "FortyTwo!",
         }
 
+    def test_update_unknown_field(self):
+        config = EYConfExtraFields(Config42())
+
+        config.update({"unknown_field": "I am unknown!"})
+
+        assert config.data.unknown_field == "I am unknown!"  # type: ignore[attr]
+        assert config._extra_data.unknown_field == "I am unknown!"
+        with pytest.raises(AttributeError):
+            _ = config._data.non_existent_field  # type: ignore[attr]
+
     def test_update_dict(self):
         @dataclass
         class SchemaDict: